+++ src/components/Cards.jsx
@@ -1,14 +1,3 @@
-// const Cards = () => {
-//   return (
-//     <div className="birds">
-//       <div className="cards">
-//         <h1>Card Component</h1>
-//       </div>
-//     </div>
-//   );
-// };
-
-// export default Cards;
 
 function Cards ({ birdData, addToCartProp }) { // name of prop is addToCart
     return (
